backend/requirements.txt
fastapi
uvicorn
python-docx
pytesseract
pillow
PyMuPDF
sentence-transformers
torch
scikit-learn